updated professor spreadsheet
