+++ backend/main.go
@@ -4,6 +4,7 @@ import (
 	"log"
 	"os"
 
+	"backend/controllers"
 	"backend/models"
 	"backend/routes"
 
@@ -24,7 +25,7 @@ func main() {
 	router := routes.SetupRouter()
 
 	// initialize the tables
-	// controllers.InitializeTables(models.DB)
+	controllers.InitializeTables(models.DB)
 
 	// Get port from environment variable or default to 3000
 	port := os.Getenv("PORT")
