+++ db.sql
@@ -1,37 +1,16 @@
--- Create the role if it doesn't exist
-DO $$
-BEGIN
-    IF NOT EXISTS (SELECT 1 FROM pg_roles WHERE rolname = 'match_rw') THEN
-        CREATE ROLE match_rw;
-    END IF;
-END $$;
+-- Create the role
+CREATE ROLE match_rw;
 
--- Create the user if it doesn't exist
-DO $$
-BEGIN
-    IF NOT EXISTS (SELECT 1 FROM pg_catalog.pg_user WHERE usename = 'match_user') THEN
-        CREATE USER match_user WITH PASSWORD 'swifty';
-    END IF;
-END $$;
+-- Create the user
+CREATE USER match_user WITH PASSWORD 'swifty';
 
--- Grant the role to the user if it's not already granted
-DO $$
-BEGIN
-    IF NOT EXISTS (SELECT 1 FROM pg_catalog.pg_user u
-                   JOIN pg_catalog.pg_auth_members am ON u.usesysid = am.member
-                   JOIN pg_catalog.pg_roles r ON am.roleid = r.oid
-                   WHERE u.usename = 'match_user' AND r.rolname = 'match_rw') THEN
-        GRANT match_rw TO match_user;
-    END IF;
-END $$;
+-- Grant the role to the user
+GRANT match_rw TO match_user;
 
--- Create the database if it doesn't exist
-DO $$
-BEGIN
-    IF NOT EXISTS (SELECT 1 FROM pg_catalog.pg_database WHERE datname = 'match_db') THEN
-        CREATE DATABASE match_db WITH OWNER = match_rw;
-    END IF;
-END $$;
+
+-- Create the database
+CREATE DATABASE match_db WITH OWNER = match_rw;
+
 
 -- Switch to the match_db database
 \connect match_db;
@@ -43,29 +22,29 @@ CREATE SCHEMA IF NOT EXISTS match_schema AUTHORIZATION match_rw;
 \connect match_db match_user;
 
 -- User Table
-CREATE TABLE IF NOT EXISTS match_schema.user (
+CREATE TABLE IF NOT EXISTS match_schema.users (
     id SERIAL PRIMARY KEY,
     name VARCHAR,
     email VARCHAR
 );
 
 -- User Role Table
-CREATE TABLE IF NOT EXISTS match_schema.role (
+CREATE TABLE IF NOT EXISTS match_schema.roles (
     user_id SERIAL PRIMARY KEY,
     role VARCHAR NOT NULL
 );
 
 -- Authentication Table
-CREATE TABLE IF NOT EXISTS match_schema.auth (
+CREATE TABLE IF NOT EXISTS match_schema.auths (
     id SERIAL PRIMARY KEY,
-    user_id INT NOT NULL REFERENCES match_schema.role(user_id),
+    user_id INT NOT NULL REFERENCES match_schema.roles(user_id),
     start_time TIMESTAMP NOT NULL,
     end_time TIMESTAMP,
     auth_token VARCHAR NOT NULL UNIQUE
 );
 
 -- Course Table
-CREATE TABLE IF NOT EXISTS match_schema.course (
+CREATE TABLE IF NOT EXISTS match_schema.courses (
     id SERIAL PRIMARY KEY,
     number VARCHAR NOT NULL,
     name VARCHAR NOT NULL,
@@ -74,9 +53,9 @@ CREATE TABLE IF NOT EXISTS match_schema.course (
 );
 
 -- Course Semester Table
-CREATE TABLE IF NOT EXISTS match_schema.course_sem (
+CREATE TABLE IF NOT EXISTS match_schema.course_sems (
     id SERIAL PRIMARY KEY,
-    course_id INT NOT NULL REFERENCES match_schema.course(id),
+    course_id INT NOT NULL REFERENCES match_schema.courses(id),
     semester VARCHAR NOT NULL,
     mandatory_level VARCHAR,
     timeslot VARCHAR
@@ -92,7 +71,7 @@ CREATE TABLE IF NOT EXISTS match_schema.preferences (
 );
 
 -- Matching Iteration Table
-CREATE TABLE IF NOT EXISTS match_schema.matching_iteration (
+CREATE TABLE IF NOT EXISTS match_schema.matching_iterations (
     id SERIAL PRIMARY KEY,
     triggered_by INT,
     status VARCHAR NOT NULL,
@@ -101,11 +80,11 @@ CREATE TABLE IF NOT EXISTS match_schema.matching_iteration (
 );
 
 -- Matching Table
-CREATE TABLE IF NOT EXISTS match_schema.matching (
+CREATE TABLE IF NOT EXISTS match_schema.matchings (
     id SERIAL PRIMARY KEY,
-    matching_iteration_id INT NOT NULL REFERENCES match_schema.matching_iteration(id),
-    user_id INT NOT NULL REFERENCES match_schema.role(user_id),
-    course_sem_id INT NOT NULL REFERENCES match_schema.course_sem(id),
+    matching_iteration_id INT NOT NULL REFERENCES match_schema.matching_iterations(id),
+    user_id INT NOT NULL REFERENCES match_schema.roles(user_id),
+    course_sem_id INT NOT NULL REFERENCES match_schema.course_sems(id),
     score DECIMAL NOT NULL
 );
