+++ frontend/src/components/dragdrop/Card.tsx
