+++ frontend/src/types/faculty.ts
@@ -1,0 +1,9 @@
+export interface Faculty {
+    ID: number;
+    Name: string;
+    Email: string;
+  }
+  
+  export interface FacultyResponse {
+    faculty: Faculty[];
+  }
